+++ backend/requirements.txt
@@ -6,6 +6,7 @@ alembic==1.13.2
 annotated-types==0.7.0
 anyio==4.4.0
 asyncpg==0.29.0
+asyncstdlib==3.12.4
 attrs==23.2.0
 beautifulsoup4==4.12.3
 boto3==1.34.149
